requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
cloudinary==1.29.0
dj-database-url==0.5.0
Django==3.2.14
django-admin-interface==0.19.1
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-filter==22.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-heroku==0.3.1
django-suit==2.0a1
django-utils-six==2.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
Markdown==3.4.1
Pillow==9.2.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.4.0
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
twilio==7.12.0
typing-extensions==4.3.0
urllib3==1.26.10
validate-email==1.3
whitenoise==6.2.0
zipp==3.8.1
